# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beachcomber Api Documentation",
        "description": "How to use Beachcomber Api",
        "contact": {
            "name": "Beachcomber",
            "email": "mauricia@beachcomber.com"
        },
        "license": {
            "name": "Terms and Condition",
            "url": "https://www.beachcomber-hotels.com/fr/terms-conditions"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/{lang}/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "en"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_email",
                                    "user_password"
                                ],
                                "properties": {
                                    "user_email": {
                                        "type": "email",
                                        "example": "example@beachcomber.com"
                                    },
                                    "user_password": {
                                        "type": "password",
                                        "example": "beachcomber"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_email",
                                    "user_password"
                                ],
                                "properties": {
                                    "user_email": {
                                        "type": "email",
                                        "example": "example@beachcomber.com"
                                    },
                                    "user_password": {
                                        "type": "password",
                                        "example": "beachcomber"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "360"
                                        },
                                        "user": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errNo": {
                                            "type": "integer",
                                            "example": "number"
                                        },
                                        "errMsg": {
                                            "type": "string",
                                            "example": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errNo": {
                                            "type": "integer",
                                            "example": "number"
                                        },
                                        "errMsg": {
                                            "type": "string",
                                            "example": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errNo": {
                                            "type": "integer",
                                            "example": "number"
                                        },
                                        "errMsg": {
                                            "type": "string",
                                            "example": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}